ops_hub/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
tqdm>=4.65.0
requests>=2.31.0
